% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_pro.R
\name{aes_pro}
\alias{aes_pro}
\title{Programmatically Friendly, Standard Evaluated aes() Function Alias}
\usage{
aes_pro(...)
}
\arguments{
\item{...}{arguments to be passed to \link[ggplot2]{aes},
but passed as formulas rather than non-standard evaluated dark magic. \cr}
}
\value{
See \link[ggplot2]{aes}.
}
\description{
Programmatically friendly version of \code{ggplot2""}\link[ggplot2]{aes} function.
This function is programmatically friendly because it uses
proper standard evaluation,
instead of non-standard evaluation,
tidy evaluation,
or similar programmatically unfriendly evaluations. \cr
}
\details{
Non-Standard Evaluation is quite controversial (see \link{subsets_NSE}). \cr
\cr
Often standard-evaluated alternatives provided. \cr
But in the case of the \code{aes()} function in 'ggplot2',
the standard-evaluated alternative changes frequently,
and the ones provided so far are rather clumsy. \cr
\cr
The \code{aes_pro()} function is the standard evaluated alternative.
Due to the way \code{aes_pro()} is programmed,
it should work no matter how many times the standard evaluation techniques
change in 'ggplot2'. \cr
It should also work in older and newer versions of 'ggplot2'. \cr
To support functions in combinations with references of the variables,
the input used here are formula inputs, rather than character inputs. \cr
See Examples section below.
}
\examples{
\dontshow{if (requireNamespace("ggplot2")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
requireNamespace("ggplot2")


data("starwars", package = "dplyr")
x <- ~ sqrt(mass)
y <- ~ height
color <- ~ sex

ggplot2::ggplot(starwars, aes_pro(x, y, color = color)) +
  ggplot2::geom_point()
\dontshow{\}) # examplesIf}
}
