% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_utilities.R
\name{indx_x}
\alias{indx_x}
\alias{indx_rm}
\title{Exported Utilities}
\usage{
indx_x(i, x, xnames, xsize)

indx_rm(i, x, xnames, xsize)
}
\arguments{
\item{i}{any of the following:
\itemize{
\item a vector of length 0, in which case an empty object is returned.
\item a strictly positive vector with indices
(duplicates are allowed, resulting in duplicated indices).
\item logical vector (without \code{NA}s) of the same length as \code{x} giving the indices to select.
\item a character vector of index names
(duplicates are allowed, resulting in duplicated indices).
If an object has multiple indices with the given name,
all the corresponding indices will be selected/duplicated/re-arranged.
\item function that returns a logical vector giving the element indices to select.
}}

\item{x}{a vector, vector-like object, factor, data.frame, data.frame-like object, or a list.}

\item{xnames}{names or dimension names}

\item{xsize}{length or dimension size}
}
\value{
The subsetted object.
}
\description{
Exported utilities
}
\examples{
x <- 1:10
names(x) <- letters[1:10]
indx_x(1:5, x, names(x), length(x))
indx_rm(1:5, x, names(x), length(x))


}
