% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa2_subsets_indx_args.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{first}
\alias{last}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:last]{first}}, \code{\link[data.table]{last}}}
}}

