% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_pro.R
\name{aes_pro}
\alias{aes_pro}
\title{Programmatically Friendly, Standard Evaluated aes() Function Alias}
\usage{
aes_pro(...)
}
\arguments{
\item{...}{arguments to be passed to \link[ggplot2]{aes},
except \code{aes_pro()} forces programmatically friendly standard evaluation.}
}
\value{
See \link[ggplot2]{aes}.
}
\description{
Programmatically friendly version of 'ggplot2''s \link[ggplot2]{aes} function.
This function is programmatically friendly because it uses
proper standard evaluation,
instead of non-standard evaluation,
tidy evaluation,
or similar programmatically unfriendly evaluations. \cr
}
\details{
Non-Standard Evaluation (sometimes abbreviated as "NSE"),
is quite controversial. \cr
Consider the following example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{aplot <- "ggplot2"
library(aplot)
}\if{html}{\out{</div>}}

What package will be attached? It will not be 'ggplot2',
nor will an error occur.
Instead, the package 'aplot' will be attached. \cr
This is due to evaluating the expression 'aplot' as a quoted expression,
instead of evaluating the contents (i.e. string or formula) of the variable.
In other words: Non-Standard Evaluation. \cr
\cr
Often standard-evaluated alternatives are also provided.
But in the case of the \code{aes()} function in 'ggplot2',
the standard-evaluated alternative changes frequently,
the ones provided so far are rather clumsy. \cr
\cr
The \code{aes_pro()} function is the standard evaluated alternative.
Due to the way \code{aes_pro()} is programmed,
it should work no matter how many times the standard evaluation techniques
change in 'ggplot2'.
It should also work in older and newer versions of 'ggplot2'.
}
\examples{
\dontshow{if (requireNamespace("ggplot2")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
requireNamespace("ggplot2")


data("starwars", package = "dplyr")
x <- "mass"
y <- "height"
color <- "sex"

ggplot2::ggplot(starwars, aes_pro(x, y, color = color)) +
  ggplot2::geom_point()
\dontshow{\}) # examplesIf}
}
