% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_set.R
\name{sb_set}
\alias{sb_set}
\alias{sb_set.default}
\alias{sb_set.matrix}
\alias{sb_set.array}
\alias{sb_set.list}
\alias{sb_set.data.frame}
\title{Method to Modify Subsets of an Object By Reference}
\usage{
sb_set(x, ...)

\method{sb_set}{default}(x, i, ..., rp, tf)

\method{sb_set}{matrix}(x, row = NULL, col = NULL, i = NULL, ..., rp, tf)

\method{sb_set}{array}(x, idx = NULL, dims = NULL, rcl = NULL, i = NULL, ..., rp, tf)

\method{sb_set}{list}(x, i, ..., rp, tf)

\method{sb_set}{data.frame}(x, row = NULL, col = NULL, filter = NULL, vars = NULL, ..., rp, tf)
}
\arguments{
\item{x}{see \link{subsets_classes}.}

\item{...}{further arguments passed to or from other methods.}

\item{i, row, col, idx, dims, rcl, filter, vars}{See \link{subsets_indx_args}. \cr
An empty index selection returns the original object unchanged. \cr}

\item{rp}{an object of somewhat the same type as the selected subset of \code{x},
and the same same length as the selected subset of \code{x} or a length of 1.}

\item{tf}{the transformation function.}
}
\value{
Returns: VOID. This method modifies the object by REFERENCE. \cr
Do NOT use assignment like \code{x <- sb_set(x, ...)}. \cr
Since this function returns void, you'll just get NULL. \cr \cr
}
\description{
This is an S3 Method to replace or transform a subset of an object
\bold{By Reference}. \cr \cr
}
\details{
\bold{Transform or Replace} \cr
Specifying argument \code{tf} will transform the subset.
Specifying \code{rp} will replace the subset.
One cannot specify both \code{tf} and \code{rp}. It's either one set or the other. \cr
Note that there is not \code{sb_set()} method for factors: this is intentional. \cr
\cr
}
\section{Reference Semantics}{


The \code{sb_set()} method modifies an object \bold{by reference}. \cr
The advantage of this is that less memory is required to modify objects. \cr
But modifying an object by reference does have 3 potential disadvantages. \cr
\cr
First, the coercion rules are slightly different: see \link{subsets_classes}. \cr
\cr
Second, if 2 or more variables refer to exactly the same object,
changing one variable also changes the other ones. \cr
I.e. the following code,

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- y <- 1:16
sb_set(x, i = 1:6, rp = 8)
}\if{html}{\out{</div>}}

modifies not just \code{x}, but also \code{y}. \cr
\cr
Third, the second consequence is true even if one of the variables is locked
(see \link[base]{bindingIsLocked}). \cr
I.e. the following code,

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tinycodet::import_LL("tinycodet", "\%<-c\%")
x <- 1:16
y \%<-c\% x
sb_set(x, i = 1:6, rp = 8)
}\if{html}{\out{</div>}}

modifies both \code{x} and \code{y} without error,
even though \code{y} is a locked constant. \cr \cr
}

\examples{


# atomic objects ====

gen_mat <- function() {
  obj <- matrix(1:16, ncol = 4)
  colnames(obj) <- c("a", "b", "c", "a")
  return(obj)
}

obj <- obj2 <- gen_mat()
obj
sb_set(obj, 1:3, 1:3, rp = -1:-9)
obj2
obj <- obj2 <- gen_mat()
obj
sb_set(obj, i = \(x)x<=5, rp = -1:-5)
obj2
obj <- obj2 <- gen_mat()
obj
sb_set(obj, col = "a", rp = cbind(-1:-4, -5:-8))
obj2

obj <- obj2 <- gen_mat()
obj
sb_set(obj, 1:3, 1:3, tf = \(x) -x)
obj2
obj <- obj2 <- gen_mat()
obj
sb_set(obj, i = \(x)x<=5, tf = \(x) -x)
obj2
obj <- obj2 <- gen_mat()
obj
sb_set(obj, col = "a", tf = \(x) -x)
obj2


gen_array <- function() {
  array(1:64, c(4,4,3))
}
obj <- gen_array()
obj
sb_set(obj, list(1:3, 1:2, c(1, 3)), 1:3, rp = -1:-12)
obj
obj <- gen_array()
obj
sb_set(obj, i = \(x)x<=5, rp = -1:-5)
obj


#############################################################################

# data.frame ====

obj <- data.frame(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
str(obj) # notice that columns "a" and "c" are INTEGER (`int`)
sb_set(
  obj, filter = ~ (a >= 2) & (c <= 17), vars = is.numeric,
  tf = sqrt # WARNING: sqrt() results in `dbl`, but columns are `int`, so decimals lost
)
print(obj)

obj <- data.frame(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
obj <- sb_coe(obj, vars = is.numeric, v = as.numeric)
str(obj)
sb_set(obj,
  filter = ~ (a >= 2) & (c <= 17), vars = is.numeric,
  tf = sqrt # SAFE: coercion performed by sb_coe(); so no warnings
) 
print(obj)

obj <- data.frame(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
str(obj) # notice that columns "a" and "c" are INTEGER (`int`)
sb_set(
  obj, vars = is.numeric,
  tf = sqrt # SAFE: row=NULL & filter = NULL, so coercion performed
)
str(obj)

}
